+++ src/Interfaces/RoutesInterface.php
@@ -1,6 +1,9 @@
 <?php
-namespace GCWorld\Routing;
+namespace GCWorld\Routing\Interfaces;
 
+/**
+ * Interface RoutesInterface
+ */
 interface RoutesInterface
 {
     /**