+++ src/Processor.php
@@ -5,7 +5,6 @@ use Riimu\Kit\PHPEncoder\PHPEncoder;
 
 /**
  * Class Processor
- * @package GCWorld\Routing
  */
 class Processor
 {
@@ -200,7 +199,7 @@ class Processor
         $php = "<?php\n";
         $php .= "namespace GCWorld\\Routing\\Generated;\n";
         $php .= "\n";
-        $php .= "class MasterRoute_".self::cleanClassName($master)." Implements \\GCWorld\\Routing\\RoutesInterface\n";
+        $php .= "class MasterRoute_".self::cleanClassName($master)." Implements \\GCWorld\\Routing\\Interfaces\\RoutesInterface\n";
         $php .= "{\n";
 
         //Get File Time Function
@@ -274,7 +273,7 @@ class Processor
         $php = "<?php\n";
         $php .= "namespace GCWorld\\Routing\\Generated;\n";
         $php .= "\n";
-        $php .= "class MasterRoute_MISC Implements \\GCWorld\\Routing\\RoutesInterface\n";
+        $php .= "class MasterRoute_MISC Implements \\GCWorld\\Routing\\Interfaces\\RoutesInterface\n";
         $php .= "{\n";
 
         //Get File Time Function