+++ CHANGELOG.md
@@ -6,6 +6,24 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+## [4.4.2](https://github.com/KongHack/Routing/releases/tag/4.4.2)
+ - @GameCharmer Add Calling Method method
+
+
+
+## [4.4.1](https://github.com/KongHack/Routing/releases/tag/4.4.1)
+ - @GameCharmer Remove all compacts for PHP 7.3
+
+
+
+## [4.4.0](https://github.com/KongHack/Routing/releases/tag/4.4.0)
+ - @GameCharmer Migrated interfaces from old interfaces package to this package
+ - New Interface for JSON only handling and accompanying abstract
+ - Moved existing interfaces into an interfaces sub-namespace
+ - gcworld/interfaces updated from 3.2.0 to 3.3.1  
+   See changes: https://github.com/KongHack/Interfaces/compare/3.2.0...3.3.1  
+   Release notes: https://github.com/KongHack/Interfaces/releases/tag/3.3.1
+
 
 ## [4.3.3](https://github.com/KongHack/Routing/releases/tag/4.3.3)
  - @GameCharmer Add UUID Token